requirements.txt
# Sector Rotation Theory — Dependencies
# Install with: pip install -r requirements.txt

# --- Data Sources ---
yfinance>=0.2.18
fredapi>=0.5.0

# --- Core Data / Numerical ---
pandas>=2.0
numpy>=1.24
pyarrow>=12.0          # Parquet read/write

# --- Visualization ---
matplotlib>=3.7
seaborn>=0.12
plotly>=5.15

# --- Machine Learning ---
scikit-learn>=1.3

# --- Econometrics ---
statsmodels>=0.14

# --- Dashboard ---
streamlit>=1.28

# --- Utilities ---
python-dotenv>=1.0
